requirements.txt
requests==2.26.0
beautifulsoup4==4.12.3
pymysql==1.0.2
m3u8==3.5.0
m3u8-To-MP4==0.1.11
mega.py==1.0.8
pyppeteer==1.0.2
youtube_dl==2021.12.17
cloudscraper==1.2.71
browsermob-proxy==0.8.0
selenium==3.141.0
websocket-client==1.8.0
PyYAML==6.0.2
uuid7==0.1.0
celery==5.2.7
redis==4.5.5
flower==2.0.1
psutil==5.9.5
dropbox==11.36.2
pcloud==1.2
typing-extensions==4.12.2
sqlalchemy==2.0.36
lxml==5.3.0
jsonpath==0.82.2
scrapy-selenium==0.0.7
python-dotenv==1.0.1
webdriver-manager==4.0.2


